requirements.txt
mysql==0.0.3
mysql-connector-python==8.4.0
mysqlclient==2.2.4
pycryptodome==3.20.0
Flask==3.0.3
Flask-Cors==4.0.1
reportlab==4.2.0
requests==2.31.0
waitress==3.0.0